refactored for conciseness and readability (and for a typesafer future): - added generics - converted loops into enhanced for-loops - removed unnecessary boxing/unboxing - removed redundant type casts - replaced checks of collection emptiness using size() with isEmpty() - removed unused imports - removed redundant field initializers - minor formatting
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinx4@9316 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ src/sphinx4/edu/cmu/sphinx/frontend/frequencywarp/PLPCepstrumProducer.java
@@ -84,8 +84,7 @@ public class PLPCepstrumProducer extends BaseDataProcessor {
             double frequency = 2 * Math.PI * (double) i / period;
 
             for (int j = 0; j < numberPLPFilters; j++) {
-                cosine[i][j] = Math.cos
-                        ((double) (frequency * (j + 0.5)));
+                cosine[i][j] = Math.cos(frequency * (j + 0.5));
             }
         }
     }
